Continue splitting Menu.cpp into multiple files.
