+++ Sources/SeriousSam/GUI/Menus/Menu_starters.h
@@ -1,0 +1,67 @@
+/* Copyright (c) 2002-2012 Croteam Ltd. All rights reserved. */
+
+#ifndef SE_INCL_MENU_STARTERS_H
+#define SE_INCL_MENU_STARTERS_H
+#ifdef PRAGMA_ONCE
+  #pragma once
+#endif
+
+
+void StartVideoOptionsMenu(void);
+void StartAudioOptionsMenu(void);
+void StartNetworkMenu(void);
+void StartNetworkJoinMenu(void);
+void StartNetworkStartMenu(void);
+void StartNetworkOpenMenu(void);
+void StartSplitScreenMenu(void);
+void StartSplitStartMenu(void);
+void StartSinglePlayerNewMenuCustom(void);
+void StartSinglePlayerNewMenu(void);
+void StartSinglePlayerQuickLoadMenu(void);
+void StartSinglePlayerLoadMenu(void);
+void StartSinglePlayerSaveMenu(void);
+void StartDemoLoadMenu(void);
+void StartDemoSaveMenu(void);
+void StartNetworkQuickLoadMenu(void);
+void StartNetworkLoadMenu(void);
+void StartNetworkSaveMenu(void);
+void StartSplitScreenQuickLoadMenu(void);
+void StartSplitScreenLoadMenu(void);
+void StartSplitScreenSaveMenu(void);
+void StartVarGameOptions(void);
+void StartSinglePlayerGameOptions(void);
+void StartGameOptionsFromNetwork(void);
+void StartGameOptionsFromSplitScreen(void);
+void StartRenderingOptionsMenu(void);
+void StartCustomizeKeyboardMenu(void);
+void StartCustomizeAxisMenu(void);
+void StartOptionsMenu(void);
+void StartCurrentLoadMenu();
+void StartCurrentSaveMenu();
+void StartCurrentQuickLoadMenu();
+void StartChangePlayerMenuFromOptions(void);
+void StartChangePlayerMenuFromSinglePlayer(void);
+void StartControlsMenuFromPlayer(void);
+void StartControlsMenuFromOptions(void);
+void StartSelectLevelFromSingle(void);
+void StartHighScoreMenu(void);
+void StartSelectPlayersMenuFromSplit(void);
+void StartSelectPlayersMenuFromNetwork(void);
+void StartSelectPlayersMenuFromOpen(void);
+void StartSelectPlayersMenuFromServers(void);
+void StartSelectServerLAN(void);
+void StartSelectServerNET(void);
+void StartSelectLevelFromSplit(void);
+void StartSelectLevelFromNetwork(void);
+void StartSelectPlayersMenuFromSplitScreen(void);
+void StartSelectPlayersMenuFromNetworkLoad(void);
+void StartSelectPlayersMenuFromSplitScreenLoad(void);
+void StartPlayerModelLoadMenu(void);
+void StartControlsLoadMenu(void);
+void StartCustomLoadMenu(void);
+void StartAddonsLoadMenu(void);
+void StartModsLoadMenu(void);
+void StartNetworkSettingsMenu(void);
+void StartSinglePlayerMenu(void);
+
+#endif  /* include-once check. */
